+++ test/steps/fetch-mapped-metadata.test.js
@@ -11,7 +11,7 @@
  */
 /* eslint-env mocha */
 import assert from 'assert';
-import { PipelineStatusError } from '../../src/index.js';
+import { PipelineResponse, PipelineStatusError } from '../../src/index.js';
 import { StaticS3Loader } from '../StaticS3Loader.js';
 import fetchMappedMetadata from '../../src/steps/fetch-mapped-metadata.js';
 import { FileS3Loader } from '../FileS3Loader.js';
@@ -24,13 +24,14 @@ describe('Fetch Mapped Metadata', () => {
       contentBusId: 'foo-id',
       partition: 'live',
       mapped: true,
+      mappedPath: '/mapped',
       info: {
         path: '/mapped',
       },
       s3Loader: new FileS3Loader()
         .rewrite('foo-id/live/mapped/metadata.json', 'metadata-kv.json'),
     };
-    await fetchMappedMetadata(state);
+    await fetchMappedMetadata(state, new PipelineResponse());
     assert.deepEqual(state.mappedMetadata.getModifiers('/new/foo'), {
       description: 'Lorem ipsum dolor sit amet.',
       keywords: 'ACME, CORP, PR',
@@ -44,6 +45,7 @@ describe('Fetch Mapped Metadata', () => {
       contentBusId: 'foo-id',
       partition: 'live',
       mapped: true,
+      mappedPath: '/mapped',
       info: {
         path: '/mapped',
       },
@@ -53,7 +55,7 @@ describe('Fetch Mapped Metadata', () => {
           body: 'this is no json!',
           headers: new Map(),
         }),
-    });
+    }, new PipelineResponse());
     await assert.rejects(promise, new PipelineStatusError(500, 'failed parsing of /mapped/metadata.json: Unexpected token \'h\', "this is no json!" is not valid JSON'));
   });
 
@@ -63,6 +65,7 @@ describe('Fetch Mapped Metadata', () => {
       contentBusId: 'foo-id',
       partition: 'live',
       mapped: true,
+      mappedPath: '/mapped',
       info: {
         path: '/mapped',
       },
@@ -74,13 +77,14 @@ describe('Fetch Mapped Metadata', () => {
         }),
     });
     await assert.rejects(promise, new PipelineStatusError(500, 'failed loading of /mapped/metadata.json: data must be an array'));
-  });
+  }, new PipelineResponse());
 
   it('ignores metadata with no data array', async () => {
     const state = {
       log: console,
       contentBusId: 'foo-id',
       partition: 'live',
+      mappedPath: '/mapped',
       mapped: true,
       info: {
         path: '/mapped',
@@ -94,7 +98,7 @@ describe('Fetch Mapped Metadata', () => {
     };
     await fetchMappedMetadata(state);
     assert.strictEqual(state.mappedMetadata, Modifiers.EMPTY);
-  });
+  }, new PipelineResponse());
 
   it('throws error on generic error', async () => {
     const promise = fetchMappedMetadata({
@@ -102,6 +106,7 @@ describe('Fetch Mapped Metadata', () => {
       contentBusId: 'foo-id',
       partition: 'live',
       mapped: true,
+      mappedPath: '/mapped',
       info: {
         path: '/mapped',
       },
@@ -113,5 +118,5 @@ describe('Fetch Mapped Metadata', () => {
         }),
     });
     await assert.rejects(promise, new PipelineStatusError(502, 'failed to load /mapped/metadata.json: 500'));
-  });
+  }, new PipelineResponse());
 });